refactor: move fs API to separate module and add wasm shim
- Move all filesystem operations from project.rs to fs.rs
- Create new Fs struct with async and sync fs methods
- Add wasm_shim.rs with getWasmModule/getWasmMemory exports
- Update TypeScript code to use Fs instead of Project for fs ops
- Update loaderWorkerPool to use getWasmModule()/getWasmMemory()
- Upgrade wasm-bindgen to 0.2.108 and update CI workflows
