+++ .github/workflows/utooweb-ci.yml
@@ -43,7 +43,7 @@ jobs:
       - name: Install wasm-bindgen-cli
         uses: jetli/wasm-bindgen-action@v0.2.0
         with:
-          version: '0.2.106'
+          version: '0.2.108'
       - name: Build WASM
         run: |
           npm run build --workspace=@utoo/pack-shared
