+++ .github/workflows/utooweb-release.yml
@@ -42,7 +42,7 @@ jobs:
       - name: Install wasm-bindgen-cli
         uses: jetli/wasm-bindgen-action@v0.2.0
         with:
-          version: '0.2.106'
+          version: '0.2.108'
       - name: Build
         run: npx turbo run build --filter=@utoo/web
       - name: Get version from release or use alpha version
