+++ crates/utoo-wasm/Cargo.toml
@@ -43,8 +43,8 @@ tracing                  = { workspace = true }
 tracing-subscriber       = { workspace = true }
 tracing-web              = "0.1.3"
 utoo-ruborist            = { path = "../ruborist" }
-wasm-bindgen             = "0.2.106"
-wasm-bindgen-futures     = "0.4.56"
+wasm-bindgen             = "0.2.108"
+wasm-bindgen-futures     = "0.4.58"
 
 ## turbpack specific - conditional on utoopack feature
 pack-api            = { workspace = true, optional = true }
