+++ packages/utoo-web/src/project/InternalProject.ts
@@ -10,6 +10,7 @@ import {
 } from "../types";
 import initWasm, {
   DirEntryType,
+  Fs,
   initLogFilter,
   Project as ProjectInternal,
 } from "../utoo";
@@ -55,11 +56,10 @@ class InternalEndpoint implements ProjectEndpoint {
   }
 
   async build() {
-    const binding = await this.wasmInit!;
+    await this.wasmInit!;
 
     if (this.options?.loaderWorkerUrl && !this.loaderWorkerPoolInitialized) {
       runLoaderWorkerPool(
-        binding,
         this.options.cwd,
         this.options!.loaderWorkerUrl,
         this.options?.loadersImportMap,
@@ -74,9 +74,9 @@ class InternalEndpoint implements ProjectEndpoint {
     await this.wasmInit!;
     let ret;
     if (encoding === "utf8") {
-      ret = await ProjectInternal.readToString(path);
+      ret = await Fs.readToString(path);
     } else {
-      return await ProjectInternal.read(path);
+      return await Fs.read(path);
     }
     return ret as any;
   }
@@ -88,20 +88,20 @@ class InternalEndpoint implements ProjectEndpoint {
   ) {
     await this.wasmInit!;
     if (typeof content === "string") {
-      return await ProjectInternal.writeString(path, content);
+      return await Fs.writeString(path, content);
     } else {
-      return await ProjectInternal.write(path, content);
+      return await Fs.write(path, content);
     }
   }
 
   async copyFile(src: string, dst: string) {
     await this.wasmInit!;
-    return await ProjectInternal.copyFile(src, dst);
+    return await Fs.copyFile(src, dst);
   }
 
   async stat(path: string): Promise<Stats> {
     await this.wasmInit!;
-    const metadata = await ProjectInternal.metadata(path);
+    const metadata = await Fs.metadata(path);
     const json = metadata.toJSON() as any;
     const raw: RawStats = {
       type: json.type as DirEntryType,
@@ -114,9 +114,9 @@ class InternalEndpoint implements ProjectEndpoint {
   async readdir(path: string, options?: { recursive?: boolean }) {
     await this.wasmInit!;
     const dirEntries = options?.recursive
-      ? await ProjectInternal.readDir(path)
+      ? await Fs.readDir(path)
       : // TODO: support recursive readDirAll
-        await ProjectInternal.readDir(path);
+        await Fs.readDir(path);
     const rawDirents: RawDirent[] = dirEntries.map((e: any) => {
       const dir = e.toJSON() as any;
       return {
@@ -131,21 +131,21 @@ class InternalEndpoint implements ProjectEndpoint {
   async mkdir(path: string, options?: { recursive?: boolean }) {
     await this.wasmInit!;
     if (options?.recursive) {
-      return await ProjectInternal.createDirAll(path);
+      return await Fs.createDirAll(path);
     } else {
-      return await ProjectInternal.createDir(path);
+      return await Fs.createDir(path);
     }
   }
 
   async rm(path: string, options?: { recursive?: boolean }) {
     await this.wasmInit!;
-    let metadata = (await ProjectInternal.metadata(path)).toJSON();
+    let metadata = (await Fs.metadata(path)).toJSON();
 
     switch ((metadata as any).type as DirEntryType) {
       case "file":
-        return await ProjectInternal.removeFile(path);
+        return await Fs.removeFile(path);
       case "directory":
-        return await ProjectInternal.removeDir(path, !!options?.recursive);
+        return await Fs.removeDir(path, !!options?.recursive);
       default:
         // nothing to remove now
         break;
@@ -154,7 +154,7 @@ class InternalEndpoint implements ProjectEndpoint {
 
   async rmdir(path: string, options?: { recursive?: boolean }) {
     await this.wasmInit!;
-    return await ProjectInternal.removeDir(path, !!options?.recursive);
+    return await Fs.removeDir(path, !!options?.recursive);
   }
 
   async gzip(files: PackFile[]) {
